pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "aiwalt"
version = "0.1.0"
description = "AI voice assistant for Raspberry Pi 5 — your personal Jarvis/Griot"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "azure-cognitiveservices-speech>=1.40.0",
    "pvporcupine>=3.0.0",
    "pvrecorder>=1.2.0",
    "numpy>=1.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
aiwalt = "aiwalt.main:main"

[tool.setuptools.packages.find]
include = ["aiwalt*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
